Working buffer texture with non-interpolated coordinates

+++ src/OpenGL/Enumerations/ClearMask.cs
@@ -1,6 +1,6 @@
 ﻿namespace OpenGL;
 
-public enum ClearMask : ushort
+public enum ClearMask : uint
 {
     Color = GL_COLOR_BUFFER_BIT,
     Depth = GL_DEPTH_BUFFER_BIT,
